+++ metriken-exposition/src/lib.rs
@@ -18,5 +18,8 @@ pub use parquet::{
     ParquetCompression, ParquetHistogramType, ParquetOptions, ParquetSchema, ParquetWriter,
 };
 pub use prometheus::{prometheus_text, PrometheusOptions};
-pub use snapshot::{Counter, Gauge, Histogram, Snapshot, SnapshotV1, SnapshotV2};
+pub use snapshot::{
+    Counter, Gauge, GroupSchema, GroupSnapshot, Histogram, MetricDesc, Snapshot, SnapshotV1,
+    SnapshotV2, SnapshotV3,
+};
 pub use snapshotter::{Snapshotter, SnapshotterBuilder};